# website/docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Indexarr",
    "version": "0.1.0",
    "description": "Decentralized torrent indexing with DHT crawling, content classification, gossip-based P2P sync, and contributor identity.\n\n## Authentication\n\nMost endpoints require no authentication. Admin endpoints (API key management, nuke review, debug logging) require an API key passed via the `X-Api-Key` header. The Torznab endpoint supports an optional `apikey` query parameter.\n\n## Base URL\n\nAll REST endpoints are prefixed with `/api/v1/`. The Torznab endpoint is at `/api/torznab`. GraphQL is at `/graphql`."
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development"
    }
  ],
...
  "paths": {
...
    "/api/v1/torrent/{info_hash}": {
      "get": {
        "tags": ["Torrents"],
        "summary": "Torrent detail",
        "description": "Full torrent record including files, content classification, tags, and vote summary.",
        "parameters": [
          { "name": "info_hash", "in": "path", "required": true, "schema": { "type": "string" }, "description": "40-character hex info hash" }
        ],
        "responses": {
          "200": { "description": "Torrent detail with files, content, tags, votes" },
          "404": { "description": "Torrent not found" }
        }
      }
    },
...
    "/api/v1/system/logs/recent": {
      "get": {
        "tags": ["System"],
        "summary": "Recent log entries",
        "parameters": [
          { "name": "count", "in": "query", "schema": { "type": "integer", "default": 500, "minimum": 1, "maximum": 5000 } },
          { "name": "category", "in": "query", "schema": { "type": "string" }, "description": "Filter by log category" }
        ],
        "responses": { "200": { "description": "Log entries" } }
      }
    },
...
    "/api/v1/system/logs/debug": {
      "post": {
        "tags": ["System"],
        "summary": "Toggle debug logging",
        "description": "Enable or disable debug-level logging. Requires admin authentication.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "enabled": { "type": "boolean" }
                },
                "required": ["enabled"]
              }
            }
          }
        },
        "responses": { "200": { "description": "Debug logging toggled" } }
      }
    },
    "/api/v1/system/sync/preferences": {
      "get": {
        "tags": ["System"],
        "summary": "Sync import preferences",
        "responses": { "200": { "description": "Current sync category preferences" } }
      },
      "post": {
        "tags": ["System"],
        "summary": "Set sync preferences",
        "description": "Configure which categories to import during sync. Requires admin authentication.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "import_categories": { "type": "array", "items": { "type": "string" } },
                  "sync_comments": { "type": "boolean" }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "Preferences updated" } }
      }
    },
...
  }
}